Moved drawScene to screen.h
