+++ src/main.c
@@ -18,183 +18,6 @@ typedef enum
     THEEND
 } GameState;
 
-int compareEnemyDistance(const void *a, const void *b)
-{
-    // Load in thwo collisiondata pointers
-    CollisionData *f1 = *(CollisionData **)a;
-    CollisionData *f2 = *(CollisionData **)b;
-    // Compare their distance
-    if (!f1 || !f2)
-        return -1;
-    float cmp = f1->d - f2->d;
-    if (cmp == 0.0)
-        return 0;
-    if (cmp < 0.0)
-        return 1;
-    return -1;
-}
-
-void drawScene(Player p1, CollisionData **enemyColl, int enemycount, CollisionData **wallhits, int raycount, CollisionData **projectileData, Image *floorImage, Texture2D *floorTextureBuffer, Image floorTexture, Image roofTexture)
-{
-
-    // Group all the collisiondata into one huge array
-    CollisionData **allData = malloc(sizeof(CollisionData *) * (enemycount + raycount + MAXPROJECTILES));
-    if (!allData)
-        return;
-    memcpy(allData, enemyColl, enemycount * sizeof(CollisionData *));
-    memcpy(allData + enemycount, wallhits, raycount * sizeof(CollisionData *));
-    memcpy(allData + enemycount + raycount, projectileData, MAXPROJECTILES * sizeof(CollisionData *));
-
-    // quicksort the array based on the distance of the collisions
-    qsort(allData, (enemycount + raycount + MAXPROJECTILES), sizeof(CollisionData *), compareEnemyDistance);
-
-    Color *pixels = floorImage->data; // Pointer to the Image pixel buffer
-
-    Vec2 plane = {
-        -p1.dir.y * tanf(DEG_TO_RAD(FOV / 2)),
-        p1.dir.x * tanf(DEG_TO_RAD(FOV / 2))};
-
-    float posScale = 0.0150f; // Achieved through trial and error. If the floor is moving to much in the same direction as the player. I.e moves infront of the player. This is to low and vice versa
-
-    for (int y = SCREEN_HEIGHT / 2 + 1; y < SCREEN_HEIGHT; y++)
-    {
-        float rayDirX0 = p1.dir.x - plane.x;
-        float rayDirY0 = p1.dir.y - plane.y;
-        float rayDirX1 = p1.dir.x + plane.x;
-        float rayDirY1 = p1.dir.y + plane.y;
-
-        float rowDistance = (float)SCREEN_HEIGHT / (2.0f * y - SCREEN_HEIGHT);
-        float stepX = rowDistance * (rayDirX1 - rayDirX0) / SCREEN_WIDTH;
-        float stepY = rowDistance * (rayDirY1 - rayDirY0) / SCREEN_WIDTH;
-
-        float floorX = p1.pos.x * posScale + rowDistance * rayDirX0;
-        float floorY = p1.pos.y * posScale + rowDistance * rayDirY0;
-
-        for (int x = 0; x < SCREEN_WIDTH; ++x)
-        {
-
-            float repeatScale = 1.0f; // how much world space each texture tile covers
-
-            int tx = (int)((floorX / repeatScale) * floorTexture.width) % floorTexture.width;
-            int ty = (int)((floorY / repeatScale) * floorTexture.height) % floorTexture.height;
-
-            // Ensure wrapping is safe for negative values
-            tx = (tx + floorTexture.width) % floorTexture.width;
-            ty = (ty + floorTexture.height) % floorTexture.height;
-
-            int ceilingY = SCREEN_HEIGHT - y;
-
-            // Get the floor and ceiling colors
-            Color floorColor = GetImageColor(floorTexture, tx, ty);
-            Color ceilingColor = GetImageColor(roofTexture, tx, ty);
-
-            // Set the pixels in the Image data directly (faster than DrawPixel)
-            pixels[y * SCREEN_WIDTH + x] = floorColor;
-            pixels[ceilingY * SCREEN_WIDTH + x] = ceilingColor;
-
-            floorX += stepX;
-            floorY += stepY;
-        }
-    }
-
-    // After updating the floorImage, update the floorTextureBuffer
-    UpdateTexture(*floorTextureBuffer, floorImage->data);
-
-    // Draw the modified floorImage (both floor and ceiling) to the screen
-    DrawTexture(*floorTextureBuffer, 0, 0, WHITE); // You can adjust the position here
-
-    for (int c = 0; c < (enemycount + raycount + MAXPROJECTILES); c++)
-    {
-        if (!allData[c]) // skip null data
-            continue;
-
-        switch (isnan(allData[c]->textureOffset)) // Collisions with non wall objects have textureOffset as Nan
-        {
-        case 1: // Not a wall
-        {
-            Vec2 enemyPos = allData[c]->position;
-
-            // Vector from player to enemy
-            float dx = enemyPos.x - p1.pos.x;
-            float dy = enemyPos.y - p1.pos.y;
-
-            // Inverse camera transform
-            float invDet = 1.0f / (plane.x * p1.dir.y - p1.dir.x * plane.y);
-
-            float transformX = invDet * (p1.dir.y * dx - p1.dir.x * dy);
-            float transformY = invDet * (-plane.y * dx + plane.x * dy);
-
-            if (transformY <= 0)
-                continue; // Enemy is behind the player
-
-            // Projected X position on screen
-            float enemyScreenX = (SCREEN_WIDTH / 2) * (1 + transformX / transformY);
-
-            Texture2D sprite = allData[c]->texture;
-
-            // Preserve sprite aspect ratio
-            float aspectRatio = (float)sprite.width / (float)sprite.height;
-
-            float dist = allData[c]->d;
-            float corrected = dist * allData[c]->angle;                 // Correct fisheye effect
-            float wallHeight = (TILE_SIZE * SCREEN_HEIGHT) / corrected; // Wall height based on screen size
-
-            // Sprite height scaling factor
-            float spritesScale = 24.0;
-            float spriteHeight = spritesScale * (SCREEN_HEIGHT / transformY) * 1.8f * ((float)sprite.height / 64.0); // 1.8 = tune to taste
-            float spriteWidth = spriteHeight * aspectRatio;
-
-            Rectangle src = {
-                0, 0,
-                (float)sprite.width,
-                (float)sprite.height};
-
-            Rectangle dest = {
-                enemyScreenX - spriteWidth / 2,
-                SCREEN_HEIGHT / 2 + wallHeight / 2 - spriteHeight,
-                spriteWidth,
-                spriteHeight};
-
-            DrawTexturePro(sprite, src, dest, (Vector2){0, 0}, 0.0f, WHITE);
-            break;
-        }
-        default: // A wall
-        {
-
-            float dist = allData[c]->d;
-            float corrected = dist * vectorDot(allData[c]->rayDir, p1.dir); // Correct fisheye effect
-            float wallHeight = ((TILE_SIZE * SCREEN_HEIGHT) / corrected);   // Wall height based on screen size
-
-            Texture2D texture = allData[c]->texture;
-
-            float sliceWidth = (float)SCREEN_WIDTH / NUM_RAYS;
-            float screenX = allData[c]->id * sliceWidth;
-
-            // --- Draw walls ---
-            float texX = allData[c]->textureOffset * texture.width;
-            // Source rectangle: a vertical slice of the wall texture
-            Rectangle source = {
-                texX,
-                0,
-                1,
-                (float)texture.height};
-
-            // Destination rectangle: the scaled vertical slice on screen
-            Rectangle destination = {
-                screenX, // X on screen
-                (SCREEN_HEIGHT / 2.0f) - (wallHeight / 2.0f),
-                sliceWidth, // stretches pixels in source retangel to slicewith
-                wallHeight};
-
-            DrawTexturePro(texture, source, destination, (Vector2){0, 0}, 0.0f, WHITE);
-        }
-
-        break;
-        }
-    }
-    free(allData); // Since we memcpy the only thing stored is pointers to the other pointers and thus the data itself will be freed later
-}
-
 int main(void)
 {
     InitWindow(SCREEN_WIDTH, SCREEN_HEIGHT, "Schlem on Campus");
@@ -236,7 +59,7 @@ int main(void)
 
         CollisionData **projectileData = rayShotProjectile(player, FOV, mp, mp->projectiles); // Gets projectile CollisionData
 
-        drawScene(player, enemyData, mp->enemyCount, hits, NUM_RAYS, projectileData, &floorImage, &floorTextureBuffer, floorTexture, roofTexture);
+        drawScene(&player, (void **)enemyData, mp->enemyCount, (void **)hits, NUM_RAYS, (void **)projectileData, &floorImage, &floorTextureBuffer, floorTexture, roofTexture);
 
         // Switch between the different states
         switch (gameState)
